+++ flag/service/cluster/kubernetes/kubelet/kubelet.go
@@ -2,6 +2,5 @@ package kubelet
 
 type Kubelet struct {
 	AltNames string
-	Labels   string
 	Port     string
 }