+++ service/controller/cloudconfig/types.go
@@ -2,7 +2,7 @@ package cloudconfig
 
 import (
 	providerv1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/provider/v1alpha1"
-	"github.com/giantswarm/certs"
+	"github.com/giantswarm/certs/v2/pkg/certs"
 	k8scloudconfig "github.com/giantswarm/k8scloudconfig/v7/pkg/template"
 	"github.com/giantswarm/randomkeys"
 	capzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
@@ -56,10 +56,11 @@ type IgnitionTemplateData struct {
 	AzureMachinePool *expcapzv1alpha3.AzureMachinePool
 	AzureCluster     *capzv1alpha3.AzureCluster
 	Cluster          *capiv1alpha3.Cluster
-	ClusterCerts     certs.Cluster
 	ClusterKeys      randomkeys.Cluster
 	CustomObject     providerv1alpha1.AzureConfig
 	Images           k8scloudconfig.Images
 	MachinePool      *expcapiv1alpha3.MachinePool
+	MasterCertFiles  []certs.File
 	Versions         k8scloudconfig.Versions
+	WorkerCertFiles  []certs.File
 }